Move tag and module into separate packages
