+++ src/main/java/org/cascadebot/cascadebot/commands/management/tag/TagCommand.java
@@ -3,17 +3,19 @@
  * Licensed under the MIT license.
  */
 
-package org.cascadebot.cascadebot.commands.management;
+package org.cascadebot.cascadebot.commands.management.tag;
 
-import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Member;
 import org.cascadebot.cascadebot.commandmeta.CommandContext;
+import org.cascadebot.cascadebot.commandmeta.ICommandMain;
 import org.cascadebot.cascadebot.commandmeta.ISubCommand;
+import org.cascadebot.cascadebot.commandmeta.Module;
 import org.cascadebot.cascadebot.data.objects.Tag;
-import org.cascadebot.cascadebot.messaging.MessagingObjects;
 import org.cascadebot.cascadebot.permissions.CascadePermission;
 
-public class TagRawSubCommand implements ISubCommand {
+import java.util.Set;
+
+public class TagCommand implements ICommandMain {
 
     @Override
     public void onCommand(Member sender, CommandContext context) {
@@ -24,33 +26,34 @@ public class TagRawSubCommand implements ISubCommand {
 
         String tagName = context.getArg(0).toLowerCase();
 
-        if (!context.getData().getManagement().hasTag(tagName)) {
+        if (!context.getData().getManagement().hasTag(context.getArg(0))) {
             context.getTypedMessaging().replyDanger(context.i18n("commands.tag.cannot_find_tag", tagName));
             return;
         }
 
         Tag tag = context.getData().getManagement().getTag(tagName);
-        EmbedBuilder builder = MessagingObjects.getClearThreadLocalEmbedBuilder();
-        builder.setTitle(context.i18n("words.tag") + ": " + tagName);
-        builder.setDescription("```" + tag.getContent() + "```");
-        builder.addField(context.i18n("words.category"), tag.getCategory(), true);
+        context.reply(tag.formatTag(context));
+    }
 
-        context.getTypedMessaging().replyInfo(builder);
+    @Override
+    public Module getModule() {
+        return Module.MANAGEMENT;
     }
 
     @Override
     public String command() {
-        return "raw";
-    }
-
-    @Override
-    public String parent() {
         return "tag";
     }
 
     @Override
+    public Set<ISubCommand> getSubCommands() {
+        return Set.of(new TagCreateSubCommand(), new TagDeleteSubCommand(), new TagListSubCommand(), new TagRawSubCommand(), new TagPlaceholdersSubCommand(),
+                new TagCategorySubCommand(), new TagEditSubCommand());
+    }
+
+    @Override
     public CascadePermission getPermission() {
-        return CascadePermission.of("tag.raw", false);
+        return CascadePermission.of("tag", false);
     }
 
 }
