+++ src/main/java/org/cascadebot/cascadebot/commands/management/tag/TagRawSubCommand.java
@@ -3,19 +3,17 @@
  * Licensed under the MIT license.
  */
 
-package org.cascadebot.cascadebot.commands.management;
+package org.cascadebot.cascadebot.commands.management.tag;
 
+import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Member;
 import org.cascadebot.cascadebot.commandmeta.CommandContext;
-import org.cascadebot.cascadebot.commandmeta.ICommandMain;
 import org.cascadebot.cascadebot.commandmeta.ISubCommand;
-import org.cascadebot.cascadebot.commandmeta.Module;
 import org.cascadebot.cascadebot.data.objects.Tag;
+import org.cascadebot.cascadebot.messaging.MessagingObjects;
 import org.cascadebot.cascadebot.permissions.CascadePermission;
 
-import java.util.Set;
-
-public class TagCommand implements ICommandMain {
+public class TagRawSubCommand implements ISubCommand {
 
     @Override
     public void onCommand(Member sender, CommandContext context) {
@@ -26,34 +24,33 @@ public class TagCommand implements ICommandMain {
 
         String tagName = context.getArg(0).toLowerCase();
 
-        if (!context.getData().getManagement().hasTag(context.getArg(0))) {
+        if (!context.getData().getManagement().hasTag(tagName)) {
             context.getTypedMessaging().replyDanger(context.i18n("commands.tag.cannot_find_tag", tagName));
             return;
         }
 
         Tag tag = context.getData().getManagement().getTag(tagName);
-        context.reply(tag.formatTag(context));
-    }
+        EmbedBuilder builder = MessagingObjects.getClearThreadLocalEmbedBuilder();
+        builder.setTitle(context.i18n("words.tag") + ": " + tagName);
+        builder.setDescription("```" + tag.getContent() + "```");
+        builder.addField(context.i18n("words.category"), tag.getCategory(), true);
 
-    @Override
-    public Module getModule() {
-        return Module.MANAGEMENT;
+        context.getTypedMessaging().replyInfo(builder);
     }
 
     @Override
     public String command() {
+        return "raw";
+    }
+
+    @Override
+    public String parent() {
         return "tag";
     }
 
     @Override
-    public Set<ISubCommand> getSubCommands() {
-        return Set.of(new TagCreateSubCommand(), new TagDeleteSubCommand(), new TagListSubCommand(), new TagRawSubCommand(), new TagPlaceholdersSubCommand(),
-                new TagCategorySubCommand(), new TagEditSubCommand());
-    }
-
-    @Override
     public CascadePermission getPermission() {
-        return CascadePermission.of("tag", false);
+        return CascadePermission.of("tag.raw", false);
     }
 
 }
